Refactored Login page into components
