+++ frontend/src/styles/InputField.module.css
@@ -1,0 +1,8 @@
+.input {
+    padding: 10px;
+    margin-bottom: 1em;
+    border: 1px solid #cccccc;
+    border-radius: 4px;
+    font-size: 1em;
+}
+
