+++ frontend/src/styles/Login.module.css
@@ -20,40 +20,3 @@
     margin-bottom: 1em;
     text-align: center;
   }
-  
-  .loginForm {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .loginInput {
-    padding: 10px;
-    margin-bottom: 1em;
-    border: 1px solid #cccccc;
-    border-radius: 4px;
-    font-size: 1em;
-  }
-  
-  .loginButton {
-    background-color: #5c6bc0;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 1em;
-    transition: background-color 0.3s;
-  }
-  
-  .loginButton:hover {
-    background-color: #3f51b5;
-  }
-  
-  .errorMessage {
-    color: #e53935;
-    text-align: center;
-    margin-top: -1em;
-    margin-bottom: 1em;
-    font-size: 0.9em;
-  }
-
