+++ frontend/src/styles/LoginForm.module.css
@@ -1,0 +1,28 @@
+.loginForm {
+    display: flex;
+    flex-direction: column;
+}
+
+.errorMessage {
+    color: #e53935;
+    text-align: center;
+    margin-top: -1em;
+    margin-bottom: 1em;
+    font-size: 0.9em;
+}
+
+.loginButton {
+    background-color: #5c6bc0;
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 1em;
+    transition: background-color 0.3s;
+}
+
+.loginButton:hover {
+    background-color: #3f51b5;
+}
+
